fixed handling of context menu to modify dynamic text (only for mac)
git-svn-id: file:///Users/min/Documents/svnrepros/svn_mindia/trunk@308 5f0a215f-9dc6-43d0-aa08-c0b9dc37ed8b

+++ src/dyngraphop.cpp
@@ -1478,9 +1478,9 @@ void DynText::ChangeDefaultData( double dStartTimeInMS, double dShowTimeInMS )
 void DynText::GetDefaultData( double & dStartTimeInMS, double & dShowTimeInMS ) const
 {
 	minHandle<OpItem_Base> hItem = m_aOpContainer[ c_iShowIndex ];
-	dStartTimeInMS = hItem->GetShowAtTimeInMS();
+    dStartTimeInMS = (double)hItem->GetShowAtTimeInMS();
 	hItem = m_aOpContainer[ c_iHideStopIndex ];
-	dShowTimeInMS = hItem->GetDelayInMS() - dStartTimeInMS;
+    dShowTimeInMS = (double)(hItem->GetDelayInMS() - dStartTimeInMS);
 }
 
 void DynText::Delta( double dDeltaTimeInMS )
